# Gossip store optimizations II: the gossipd parts