… cli/test/run-large-input.c
… gossipd/routing.c
… gossipd/routing.h
… gossipd/test/run-bench-find_route.c
… gossipd/test/run-find_route-specific.c
… gossipd/test/run-find_route.c
… tests/test_gossip.py
… tests/test_pay.py