@rustyrussell 2019-04-10
Fixed some bugs found by travis, and rebased since merge of prereq.